refactor: modularize core options logic and implement code splitting
- Refactored `optionPokemonData` and `optionCombat` into smaller, specialized helper functions for better maintainability.
- Replaced hardcoded constants (e.g., Unown ID 201) with dynamic helpers (`unownId()`, `struggleEnergy()`, etc.) via the Config model.
- Completed the transition to `React.lazy` and `Suspense` for all page-level components in `App.tsx`.
- Updated `App.tsx` state management to use `useState` for the stylesheet list instead of `useRef`.
- Added `tsx` dependency and updated `.gitignore` for data distribution.
- Cleaned up TypeScript interfaces and improved type safety across `assets`, `combat`, and `pokemon` options.
