+++ src/App.tsx
@@ -1,38 +1,10 @@
-import React, { createContext, useContext, useEffect, useMemo, useRef, useState } from 'react';
+import React, { createContext, lazy, Suspense, useContext, useEffect, useMemo, useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
 import './App.scss';
 
-import News from './pages/News/News';
-import Pokedex from './pages/Pokedex/Pokedex';
-import SearchPokemon from './pages/Search/Pokemon/Search';
-import SearchMove from './pages/Search/Moves/SearchMoves';
-import TypeEffect from './pages/TypeEffect/TypeEffect';
-import Weather from './pages/Weather/Weather';
-import Pokemon from './pages/Pokemon/Pokemon';
-import FindTable from './pages/Tools/FindTable/FindTable';
-import CalculateStats from './pages/Tools/CalculateStats/CalculateStats';
-import Damage from './pages/Tools/BattleDamage/Damage';
-import DpsTdo from './pages/Sheets/DpsTdo/DpsTdo';
-import Move from './pages/Move/Move';
-import Error from './pages/Error/Error';
-import Leagues from './pages/PVP/Leagues/Leagues';
-import SearchBattle from './pages/Tools/SearchBattle/SearchBattle';
-import StatsInfo from './pages/Tools/StatsInfo/StatsInfo';
-import Sticker from './pages/Sticker/Sticker';
-import RaidBattle from './pages/Tools/RaidBattle/RaidBattle';
-import CalculatePoint from './pages/Tools/CalculatePoint/CalculatePoint';
-import RankingPVP from './pages/PVP/Ranking/PVP';
-import PokemonPVP from './pages/PVP/Pokemon/Pokemon';
-import PVPHome from './pages/PVP/Home';
-import TeamPVP from './pages/PVP/Teams/PVP';
-import Battle from './pages/PVP/Battle/Battle';
-
 import Spinner from './components/Spinner/Spinner';
-import CatchChance from './pages/Tools/CatchChance/CatchChance';
 import { useLocalStorage } from 'usehooks-ts';
-import SearchTypes from './pages/Search/Types/Types';
-import StatsRanking from './pages/Sheets/StatsRanking/StatsRanking';
 import { ThemeProvider, createTheme } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import { getDesignThemes } from './utils/models/overrides/themes.model';
@@ -43,9 +15,10 @@ import { debounce } from 'lodash';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
 import { clearLocalStorageExcept } from './utils/configs/local-storage.config';
 import { getStyleList } from './utils/utils';
+import { IStyleData } from './utils/models/util.model';
 import { defaultOptions, OptionsContext } from './contexts/options.context';
 import useOptionsObserver from './utils/hooks/useOptionsObserver';
-import { loadDataDelay, transitionTime } from './utils/helpers/options-context.helpers';
+import { transitionTime } from './utils/helpers/options-context.helpers';
 import useTimestamp from './composables/useTimestamp';
 import useSpinner from './composables/useSpinner';
 import useDevice from './composables/useDevice';
@@ -53,6 +26,34 @@ import { useTheme as useThemeStore } from './composables/useTheme';
 import useRouter from './composables/useRouter';
 import ResponsiveAppBar from './components/Commons/Navbars/ResponsiveAppBar';
 import { SnackbarProvider } from './contexts/snackbar.context';
+
+const News = lazy(() => import('./pages/News/News'));
+const Pokedex = lazy(() => import('./pages/Pokedex/Pokedex'));
+const SearchPokemon = lazy(() => import('./pages/Search/Pokemon/Search'));
+const SearchMove = lazy(() => import('./pages/Search/Moves/SearchMoves'));
+const TypeEffect = lazy(() => import('./pages/TypeEffect/TypeEffect'));
+const Weather = lazy(() => import('./pages/Weather/Weather'));
+const Pokemon = lazy(() => import('./pages/Pokemon/Pokemon'));
+const FindTable = lazy(() => import('./pages/Tools/FindTable/FindTable'));
+const CalculateStats = lazy(() => import('./pages/Tools/CalculateStats/CalculateStats'));
+const Damage = lazy(() => import('./pages/Tools/BattleDamage/Damage'));
+const DpsTdo = lazy(() => import('./pages/Sheets/DpsTdo/DpsTdo'));
+const Move = lazy(() => import('./pages/Move/Move'));
+const Error = lazy(() => import('./pages/Error/Error'));
+const Leagues = lazy(() => import('./pages/PVP/Leagues/Leagues'));
+const SearchBattle = lazy(() => import('./pages/Tools/SearchBattle/SearchBattle'));
+const StatsInfo = lazy(() => import('./pages/Tools/StatsInfo/StatsInfo'));
+const Sticker = lazy(() => import('./pages/Sticker/Sticker'));
+const RaidBattle = lazy(() => import('./pages/Tools/RaidBattle/RaidBattle'));
+const CalculatePoint = lazy(() => import('./pages/Tools/CalculatePoint/CalculatePoint'));
+const RankingPVP = lazy(() => import('./pages/PVP/Ranking/PVP'));
+const PokemonPVP = lazy(() => import('./pages/PVP/Pokemon/Pokemon'));
+const PVPHome = lazy(() => import('./pages/PVP/Home'));
+const TeamPVP = lazy(() => import('./pages/PVP/Teams/PVP'));
+const Battle = lazy(() => import('./pages/PVP/Battle/Battle'));
+const CatchChance = lazy(() => import('./pages/Tools/CatchChance/CatchChance'));
+const SearchTypes = lazy(() => import('./pages/Search/Types/Types'));
+const StatsRanking = lazy(() => import('./pages/Sheets/StatsRanking/StatsRanking'));
 
 const ColorModeContext = createContext({
   toggleColorMode: () => true,
@@ -74,9 +75,13 @@ function App() {
   const [isLoaded, setIsLoaded] = useState(false);
 
   const [currentVersion, setCurrentVersion] = useState<string>();
-  const styleSheet = useRef(getStyleList());
+  const [styleSheet, setStyleSheet] = useState<IStyleData[]>([]);
 
   useOptionsObserver();
+
+  useEffect(() => {
+    setStyleSheet(getStyleList());
+  }, []);
 
   useEffect(() => {
     setTimeout(() => {
@@ -112,7 +117,6 @@ function App() {
   }, [timestampGameMaster]);
 
   useEffect(() => {
-    const controller = new AbortController();
     if (!isLoaded) {
       const currentVersion = process.env.REACT_APP_VERSION;
       setCurrentVersion(currentVersion);
@@ -121,7 +125,7 @@ function App() {
       setIsLoaded(true);
       const isCurrentVersion = currentVersion === version;
       setStateVersion(currentVersion || '');
-      loadData(controller.signal, isCurrentVersion);
+      loadTimestamp(isCurrentVersion);
     }
   }, [isLoaded]);
 
@@ -130,66 +134,41 @@ function App() {
     loadTheme(stateTheme, setStateTheme);
   }, []);
 
-  const loadData = (signal: AbortSignal, isCurrentVersion: boolean, delay = loadDataDelay()) => {
-    return new Promise<void>((resolve, reject) => {
-      const resolveHandler = async () => {
-        resolve(await loadTimestamp(isCurrentVersion));
-      };
-
-      const debouncedResolve = debounce(resolveHandler, delay);
-
-      if (signal instanceof AbortSignal) {
-        const abortHandler = () => {
-          debouncedResolve.cancel();
-          reject();
-        };
-
-        signal.addEventListener('abort', abortHandler, { once: true });
-
-        const originalResolve = debouncedResolve;
-        debouncedResolve.cancel = () => {
-          signal.removeEventListener('abort', abortHandler);
-          originalResolve.cancel();
-        };
-      }
-
-      debouncedResolve();
-    });
-  };
-
   return (
     <Box className="tw-min-h-full" sx={{ backgroundColor: 'background.default', transition: transitionTime() }}>
       <ResponsiveAppBar toggleColorMode={colorMode.toggleColorMode} version={currentVersion} />
-      <Routes>
-        <Route path="/" element={<Pokedex styleSheet={styleSheet.current} />} />
-        <Route path="/news" element={<News />} />
-        <Route path="/type-effective" element={<TypeEffect />} />
-        <Route path="/weather-boosts" element={<Weather />} />
-        <Route path="/search-pokemon" element={<SearchPokemon />} />
-        <Route path="/pokemon/:id" element={<Pokemon />} />
-        <Route path="/search-moves" element={<SearchMove />} />
-        <Route path="/move/:id" element={<Move />} />
-        <Route path="/search-types" element={<SearchTypes styleSheet={styleSheet.current} />} />
-        <Route path="/find-cp-iv" element={<FindTable />} />
-        <Route path="/calculate-stats" element={<CalculateStats />} />
-        <Route path="/search-battle-stats" element={<SearchBattle />} />
-        <Route path="/stats-table" element={<StatsInfo />} />
-        <Route path="/damage-calculate" element={<Damage />} />
-        <Route path="/raid-battle" element={<RaidBattle />} />
-        <Route path="/calculate-point" element={<CalculatePoint />} />
-        <Route path="/calculate-catch-chance" element={<CatchChance />} />
-        <Route path="/dps-tdo-sheets" element={<DpsTdo />} />
-        <Route path="/stats-ranking" element={<StatsRanking />} />
-        <Route path="/pvp" element={<PVPHome />} />
-        <Route path="/pvp/rankings/:serie/:cp" element={<RankingPVP styleSheet={styleSheet.current} />} />
-        <Route path="/pvp/teams/:serie/:cp" element={<TeamPVP styleSheet={styleSheet.current} />} />
-        <Route path="/pvp/battle" element={<Battle />} />
-        <Route path="/pvp/battle/:cp" element={<Battle />} />
-        <Route path="/pvp/:cp/:serie/:pokemon" element={<PokemonPVP styleSheet={styleSheet.current} />} />
-        <Route path="/battle-leagues" element={<Leagues />} />
-        <Route path="/stickers" element={<Sticker />} />
-        <Route path="*" element={<Error />} />
-      </Routes>
+      <Suspense>
+        <Routes>
+          <Route path="/" element={<Pokedex styleSheet={styleSheet} />} />
+          <Route path="/news" element={<News />} />
+          <Route path="/type-effective" element={<TypeEffect />} />
+          <Route path="/weather-boosts" element={<Weather />} />
+          <Route path="/search-pokemon" element={<SearchPokemon />} />
+          <Route path="/pokemon/:id" element={<Pokemon />} />
+          <Route path="/search-moves" element={<SearchMove />} />
+          <Route path="/move/:id" element={<Move />} />
+          <Route path="/search-types" element={<SearchTypes styleSheet={styleSheet} />} />
+          <Route path="/find-cp-iv" element={<FindTable />} />
+          <Route path="/calculate-stats" element={<CalculateStats />} />
+          <Route path="/search-battle-stats" element={<SearchBattle />} />
+          <Route path="/stats-table" element={<StatsInfo />} />
+          <Route path="/damage-calculate" element={<Damage />} />
+          <Route path="/raid-battle" element={<RaidBattle />} />
+          <Route path="/calculate-point" element={<CalculatePoint />} />
+          <Route path="/calculate-catch-chance" element={<CatchChance />} />
+          <Route path="/dps-tdo-sheets" element={<DpsTdo />} />
+          <Route path="/stats-ranking" element={<StatsRanking />} />
+          <Route path="/pvp" element={<PVPHome />} />
+          <Route path="/pvp/rankings/:serie/:cp" element={<RankingPVP styleSheet={styleSheet} />} />
+          <Route path="/pvp/teams/:serie/:cp" element={<TeamPVP styleSheet={styleSheet} />} />
+          <Route path="/pvp/battle" element={<Battle />} />
+          <Route path="/pvp/battle/:cp" element={<Battle />} />
+          <Route path="/pvp/:cp/:serie/:pokemon" element={<PokemonPVP styleSheet={styleSheet} />} />
+          <Route path="/battle-leagues" element={<Leagues />} />
+          <Route path="/stickers" element={<Sticker />} />
+          <Route path="*" element={<Error />} />
+        </Routes>
+      </Suspense>
       <Spinner />
     </Box>
   );
